@@ -35,8 +35,11 @@
     "grid"
   ],
   "devDependencies": {
-    "babel-core": "^5.1.13",
-    "babel-loader": "^5.0.0",
+    "autoprefixer-loader": "^3.1.0",
+    "babel-core": "^6.1.2",
+    "babel-loader": "^6.1.0",
+    "babel-preset-es2015": "^6.1.2",
+    "babel-preset-react": "^6.1.2",
     "css-loader": "^0.9.0",
     "exports-loader": "^0.6.2",
     "extract-text-webpack-plugin": "^0.3.8",
@@ -44,14 +47,14 @@
     "file-loader": "^0.8.1",
     "guid": "0.0.12",
     "gulp": "^3.8.11",
-    "gulp-babel": "^5.1.0",
-    "gulp-react": "^3.0.0",
+    "gulp-babel": "^6.1.0",
     "jsdom": "5.1.0",
     "json-loader": "^0.5.2",
     "mocha": "^2.2.1",
     "node-libs-browser": "^0.5.2",
     "parse-keys": "^1.3.0",
     "raw-loader": "^0.5.1",
+    "react-dom": "^0.14.2",
     "react-hot-loader": "^1.2.7",
     "should": "^6.0.1",
     "sorty": "^1.2.0",
@@ -79,7 +82,8 @@
     "whatwg-fetch": "^0.7.0"
   },
   "peerDependencies": {
-    "react": ">=0.13.0"
+    "react": ">=0.14.0",
+    "react-dom": ">=0.14.0"
   },
   "author": "ZippyUI <contact@zippyui.com>",
   "contributors": [],
